MLCOMPUTE-1160 | adding tests
